moved AddPlace container to Places folder, adjusted import paths accordingly
