+++ app/containers/Places/AddPlace/tests/reducer.test.js
