+++ app/containers/Places/index.js
@@ -19,7 +19,7 @@ import makeSelectPlaces from './selectors';
 import reducer from './reducer';
 import saga from './saga';
 import { listPlacesAction, deletePlacesAction, updatePlacesAction } from './actions';
-import AddPlace from 'containers/AddPlace';
+import AddPlace from './AddPlace';
 import { styles, headings } from 'assets/styles/variables';
