+++ app/i18n.js
@@ -8,7 +8,7 @@ import { addLocaleData } from 'react-intl';
 import enLocaleData from 'react-intl/locale-data/en';
 import deLocaleData from 'react-intl/locale-data/de';
 
-import { DEFAULT_LOCALE } from '../app/containers/App/constants';
+import { DEFAULT_LOCALE } from 'containers/App/constants';
 
 import enTranslationMessages from './translations/en.json';
 import deTranslationMessages from './translations/de.json';
